+++ src/Facades/Ssl.php
@@ -8,8 +8,18 @@ use Jinomial\LaravelSsl\SslManager;
 /**
  * @see \Jinomial\LaravelSsl\LaravelSsl
  */
-class Ssl extends Facade
+/**
+ * @method static \Jinomial\LaravelSsl\Support\CertificateCollection show(string|array $host, string $port = '443', array $options = [])
+ * @method static \Jinomial\LaravelSsl\Drivers\Driver driver(string|null $name = null)
+ * @method static \Jinomial\LaravelSsl\Drivers\File file()
+ * @method static \Jinomial\LaravelSsl\Drivers\OpenSsl openssl()
+ * @method static \Jinomial\LaravelSsl\Drivers\Stream stream()
+ *
+ * @see \Jinomial\LaravelSsl\SslManager
+ */
+final class Ssl extends Facade
 {
+    #[\Override]
     protected static function getFacadeAccessor()
     {
         return SslManager::class;