+++ tests/Integration/Facades/SslTest.php
@@ -3,7 +3,6 @@
 namespace Jinomial\LaravelSsl\Tests\Integration\Facades;
 
 use Jinomial\LaravelSsl\Drivers\Driver;
-use Jinomial\LaravelSsl\Drivers\OpenSsl;
 use Jinomial\LaravelSsl\Facades\Ssl;
 
 const HOST = 'jinomial.com';
@@ -14,14 +13,14 @@ uses()->group('facades');
 
 it('can show ' . HOST, function () {
     $response = Ssl::show(HOST, 443);
-    expect($response[0]['certificate']['subject']['CN'])->toEqual(HOST);
+    expect($response->first()->getCommonName())->toEqual(HOST);
 })->group('network');
 
 it('can get issuer', function () {
     $response = Ssl::show(ISSUER_HOST, 443, [
-        OpenSsl::OPTION_ID_AD_CAISSUERS => true,
+        \Jinomial\LaravelSsl\Drivers\OpenSsl::OPTION_ID_AD_CAISSUERS => true,
     ]);
-    expect($response[0]['certificate']['subject']['CN'])->toEqual(ISSUER_CN);
+    expect($response->first()->getCommonName())->toEqual(ISSUER_CN);
 })->group('network');
 
 it('can access the openssl driver by name', function () {